+++ crates/sncast/src/helpers/scarb_utils.rs
@@ -1,4 +1,4 @@
-use anyhow::{anyhow, Context, Result};
+use anyhow::{anyhow, bail, Context, Result};
 use camino::{Utf8Path, Utf8PathBuf};
 use scarb_api::{
     get_contracts_artifacts_and_source_sierra_paths,
@@ -7,6 +7,11 @@ use scarb_api::{
 };
 use scarb_ui::args::PackagesFilter;
 use shared::{command::CommandExt, print::print_as_warning};
+use starknet::{
+    core::types::{contract::SierraClass, BlockId, FlattenedSierraClass},
+    providers::{jsonrpc::HttpTransport, JsonRpcClient, Provider, ProviderError},
+};
+use starknet_crypto::FieldElement;
 use std::collections::HashMap;
 use std::env;
 use std::str::FromStr;
@@ -173,6 +178,61 @@ pub fn build_and_load_artifacts(
                 .map(|(name, (artifacts, _))| (name, artifacts))
                 .collect(),
         )
+    }
+}
+
+pub fn read_manifest_and_build_artifacts(
+    package: &Option<String>,
+    json: bool,
+    profile: &Option<String>,
+) -> Result<HashMap<String, StarknetContractArtifacts>> {
+    let manifest_path = assert_manifest_path_exists()?;
+    let package_metadata = get_package_metadata(&manifest_path, package)?;
+
+    let profile = profile.to_owned().unwrap_or("dev".to_string());
+
+    let build_config = BuildConfig {
+        scarb_toml_path: manifest_path,
+        json,
+        profile,
+    };
+
+    build_and_load_artifacts(&package_metadata, &build_config).context("Failed to build contract")
+}
+
+pub struct CompiledContract {
+    pub class: FlattenedSierraClass,
+    pub class_hash: FieldElement,
+}
+
+impl TryFrom<&StarknetContractArtifacts> for CompiledContract {
+    type Error = anyhow::Error;
+
+    fn try_from(artifacts: &StarknetContractArtifacts) -> Result<Self, Self::Error> {
+        let class = serde_json::from_str::<SierraClass>(&artifacts.sierra)
+            .context("Failed to parse sierra artifact")?
+            .flatten()?;
+
+        let class_hash = class.class_hash();
+
+        Ok(Self { class, class_hash })
+    }
+}
+
+impl CompiledContract {
+    pub async fn is_declared(&self, provider: &JsonRpcClient<HttpTransport>) -> Result<bool> {
+        let block_id = BlockId::Tag(starknet::core::types::BlockTag::Pending);
+        let class_hash = self.class_hash;
+
+        let response = provider.get_class(block_id, class_hash).await;
+
+        match response {
+            Ok(_) => Ok(true),
+            Err(ProviderError::StarknetError(
+                starknet::core::types::StarknetError::ClassHashNotFound,
+            )) => Ok(false),
+            Err(other) => bail!(other),
+        }
     }
 }